+++ yivi_app/integration_test/test_all.dart
@@ -5,6 +5,7 @@ import "disclosure_session/disclosure_session_test_all.dart"
     as disclosure_session;
 import "driving_licence_test.dart" as driving_licence_test;
 import "enroll_test.dart" as enroll_test;
+import "id_card_test.dart" as id_card_test;
 import "issuance_test.dart" as issuance_test;
 import "issue_sdjwt_over_irma_test.dart" as irma_sdjwt_test;
 import "issue_wizard_test.dart" as issue_wizard_test;
@@ -36,4 +37,5 @@ void main() {
   notifications_test.main();
   passport_test.main();
   driving_licence_test.main();
+  id_card_test.main();
 }